% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoRflow_cmip6_data_download.R
\name{geoRflow_cmip6_data_download}
\alias{geoRflow_cmip6_data_download}
\title{Download CMIP6 Climate Data}
\usage{
geoRflow_cmip6_data_download(
  model,
  timeframe,
  ensemble,
  climate_variable,
  start_year,
  end_year,
  output_folder,
  timeout
)
}
\arguments{
\item{model}{The climate model used in the dataset.}

\item{timeframe}{The timeframe for which data is being requested.}

\item{ensemble}{The ensemble identifier for the dataset.}

\item{climate_variable}{The specific climate variable required.}

\item{start_year}{The starting year for the dataset.}

\item{end_year}{The ending year for the dataset.}

\item{output_folder}{The directory where the downloaded files will be stored.}

\item{timeout}{The maximum time in seconds allowed for the download of each file.}
}
\value{
The function does not return a value but downloads files to the specified output folder.
}
\description{
This function downloads climate data from the CMIP6 dataset. It constructs a URL to
access the data based on user-specified parameters, retrieves an XML catalog, and
then downloads the files that match the specified criteria (model, timeframe,
ensemble, climate variable, and year range). The function checks for the existence
of the output directory and creates it if necessary. It handles timeouts and retries
for downloads.
}
\examples{
\donttest{
if (interactive()) {
  # Example usage of geoRflow_cmip6_data_download
  geoRflow_cmip6_data_download("exampleModel", "timeframe", "ensemble",
    "climateVariable", 1990, 2000, "outputFolder", 600)
}
}
}
