% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoRflow_cmip6_spatial_subset_download.R
\name{geoRflow_cmip6_spatial_subset_download}
\alias{geoRflow_cmip6_spatial_subset_download}
\title{Download Spatial Subset of CMIP6 Climate Data}
\usage{
geoRflow_cmip6_spatial_subset_download(
  model,
  timeframe,
  ensemble,
  climate_variable,
  start_year,
  end_year,
  north,
  south,
  east,
  west,
  output_folder,
  timeout
)
}
\arguments{
\item{model}{The climate model used in the dataset.}

\item{timeframe}{The timeframe for which data is being requested.}

\item{ensemble}{The ensemble identifier for the dataset.}

\item{climate_variable}{The specific climate variable required.}

\item{start_year}{The starting year for the dataset.}

\item{end_year}{The ending year for the dataset.}

\item{north}{The northern boundary of the spatial subset.}

\item{south}{The southern boundary of the spatial subset.}

\item{east}{The eastern boundary of the spatial subset.}

\item{west}{The western boundary of the spatial subset.}

\item{output_folder}{The directory where the downloaded files will be stored.}

\item{timeout}{The maximum time in seconds allowed for the download of each file.}
}
\value{
The function does not return a value but downloads files to the specified output folder.
}
\description{
This function downloads a spatial subset of climate data from the CMIP6 dataset based on
specified parameters. It constructs URLs for the NetCDF Subset Service (NCSS) and downloads
the files that match the specified criteria (model, timeframe, ensemble, climate variable,
year range, and spatial coordinates). The function checks for the existence of the output
directory and creates it if necessary. It handles timeouts for downloads.
}
\examples{
# Example usage of geoRflow_cmip6_spatial_subset_download
geoRflow_cmip6_spatial_subset_download("exampleModel", "timeframe", "ensemble",
  "climateVariable", 1990, 2000, 40, -40, 60, -60, "outputFolder", 600)

}
