% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoRflow_export_layers.R
\name{geoRflow_export_layers}
\alias{geoRflow_export_layers}
\title{Export Each Layer from a NetCDF File}
\usage{
geoRflow_export_layers(raster_object, layer_indices, output_dir = getwd())
}
\arguments{
\item{raster_object}{A RasterBrick or SpatRaster object.}

\item{layer_indices}{A numeric vector indicating which layers to export.}

\item{output_dir}{A string specifying the directory where the GeoTIFF files will be saved.}
}
\value{
Invisible NULL.
}
\description{
Export Each Layer from a NetCDF File
}
\examples{
\dontrun{
  # Assuming 'raster_brick' is a RasterBrick object with multiple layers
  geoRflow_export_layers(raster_brick, layer_indices = 1:10,
                         output_dir = "path/to/output")

  # If 'spat_raster' is a SpatRaster object with multiple layers
  geoRflow_export_layers(spat_raster, layer_indices = c(2, 5, 7),
                         output_dir = "path/to/different/output")

  # To export all layers from a RasterBrick object
  all_layers <- 1:nlayers(raster_brick)
  geoRflow_export_layers(raster_brick, layer_indices = all_layers,
                         output_dir = "path/to/all/layers")

  # To export layers by specifying a range
  geoRflow_export_layers(raster_brick, layer_indices = 3:8,
                         output_dir = "path/to/selected/layers")

  # To export a single layer
  geoRflow_export_layers(raster_brick, layer_indices = 5,
                         output_dir = "path/to/single/layer")

  # Assuming the working directory is the desired output directory
  geoRflow_export_layers(raster_brick, layer_indices = 1:10)

  # Using a SpatRaster object and specifying a non-existent directory
  # (it will be created)
  geoRflow_export_layers(spat_raster, layer_indices = 1:5,
                         output_dir = "path/to/new/output/dir")
}

}
