% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoRflow_join_df_vect.R
\name{geoRflow_join_df_vect}
\alias{geoRflow_join_df_vect}
\title{Spatial Join of a Data Frame with a SpatVector}
\usage{
geoRflow_join_df_vect(df, lon, lat, SpatVect)
}
\arguments{
\item{df}{A data frame to be spatially joined.}

\item{lon}{The name of the column in `df` representing longitude.}

\item{lat}{The name of the column in `df` representing latitude.}

\item{SpatVect}{A SpatVector object to join with the data frame.}
}
\value{
A data frame resulting from the spatial join of `df` with `SpatVect`.
}
\description{
Converts a data frame into a SpatVector object, performs a spatial join with
another SpatVector, and then converts the result back into a data frame. This
function is useful for integrating non-spatial data with spatial data based on
geographic coordinates.
}
\examples{
# geoRflow_join_df_vect(df, "lon", "lat", SpatVect)

}
