% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoRflow_raster_mask.R
\name{geoRflow_raster_mask}
\alias{geoRflow_raster_mask}
\title{Mask a Raster with a Vector after Reprojection}
\usage{
geoRflow_raster_mask(Spatraster, Spatvect, project_crs, projection_method)
}
\arguments{
\item{Spatraster}{A SpatRaster object representing the raster to be reprojected and masked.}

\item{Spatvect}{A SpatVector object used to mask the reprojected raster.}

\item{project_crs}{A character string or an object of class `CRS` representing
the target CRS for reprojection.}

\item{projection_method}{A character string specifying the method used for reprojection.
This should be one of the methods supported by the `terra::project` function.}
}
\value{
A SpatRaster object that is the result of masking the reprojected raster with the SpatVector.
}
\description{
This function takes a SpatRaster and a SpatVector, reprojects the raster to a
specified coordinate reference system (CRS), and then masks it using the vector.
The function checks and sets the CRS of the SpatRaster to EPSG:4326 if it's not defined,
using functions from the `terra` package. The raster is reprojected using a specified
projection method. If reprojection fails, the function stops with an error message.
}
\examples{
# Example usage
# geoRflow_raster_mask(raster_object, vector_object, "EPSG:3857", "bilinear")

}
