% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoRflow_netcdf_df.R
\name{geoRflow_netcdf_df}
\alias{geoRflow_netcdf_df}
\title{Convert Raster Layers to Data Frames}
\usage{
geoRflow_netcdf_df(raster_object, layer_indices)
}
\arguments{
\item{raster_object}{A RasterBrick or SpatRaster object.}

\item{layer_indices}{A numeric vector indicating the indices of the layers to be
converted to data frames.}
}
\value{
A list of data frames, each corresponding to a layer in the raster object.
}
\description{
This function converts specified layers of a RasterBrick or SpatRaster object
into data frames. It checks the class of the input raster object, validates the
specified layer indices, and extracts each layer as a separate data frame.
For layer names that represent dates, it attempts to format them into a standard
date format. It shows progress using a progress bar.
}
\examples{
# geoRflow_netcdf_df(raster_object, c(1, 3, 5))

}
